refactor: CommonBtn을 공통으로 사용하기 위한 디렉토리 조정
